pyproject.toml
[project]
name = "ccmonitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "colorama>=0.4.6",
    "fastapi>=0.104.0",
    "jsonschema>=4.25.0",
    "numpy>=1.24.0",
    "orjson>=3.11.1",
    "psutil>=7.0.0",
    "pydantic>=2.0.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.23.0",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0",
    "regex>=2025.7.34",
    "rich>=14.1.0",
    "scikit-learn>=1.3.0",
    "sqlite-utils>=3.38",
    "structlog>=25.4.0",
    "textual>=5.3.0",
    "textual-dev >=1.7.0",
    "tqdm>=4.66.0",
    "types-pyyaml>=6.0.12.20250809",
    "uvicorn>=0.24.0",
    "watchdog>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "types-psutil>=5.9.5",
]

[project.scripts]
ccmonitor = "src.cli.main:cli"

[tool.black]
line-length = 79
target-version = ["py311"]
skip-string-normalization = false
preview = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["ALL"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.format]
preview = true

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.pylint]
max-statements = 30

[tool.ruff.lint.mccabe]
max-complexity = 5

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "click",
    "click.*",
    "colorama",
    "colorama.*",
    "textual",
    "textual.*",
    "pytest",
    "pytest.*",
    "psutil",
    "psutil.*",
    "rich",
    "rich.*",
]
ignore_missing_imports = true
disallow_any_unimported = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--tb=short",
    "-v"
]
asyncio_mode = "auto"
markers = [
    "performance: marks tests as performance benchmarks",
    "integration: marks tests as integration tests",
    "accessibility: marks tests as accessibility compliance tests",
    "visual: marks tests as visual regression tests",
    "widget: marks tests as widget-specific tests",
    "benchmark: marks tests as benchmark tests",
]

[dependency-groups]
dev = [
    "hypothesis>=6.137.1",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "pytest-textual-snapshot>=1.1.0",
    "pytest-xdist>=3.8.0",
]
